backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pymupdf>=1.25.0
sentence-transformers>=2.7.0
python-multipart>=0.0.6
numpy>=2.0.0
pydantic>=2.5.0
aiofiles>=23.2.0
python-docx>=1.1.0
python-pptx>=0.6.23
email-validator>=2.0.0
